# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChatGPT Monitor API
  version: 0.6.0
  description: STEP-07 API contract. Authentication uses same-origin Secure cookies; credentials, OAuth device codes, channel secrets, and raw upstream evidence are never returned. Notification channels are configuration-only and remain disconnected.
  license:
    name: Proprietary
servers:
  - url: https://127.0.0.1:8080
paths:
...
  /api/auth/csrf:
    get:
      operationId: issueCSRF
      summary: Issue or rotate the double-submit CSRF token
      security: []
      responses:
        '200':
          description: CSRF token issued in JSON and a Secure SameSite cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSRFResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/password:
    post:
      operationId: verifyPassword
      summary: Verify the administrator password and issue a one-time TOTP challenge
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRequest'
      responses:
        '200':
          description: Password accepted; opaque challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '403':
          $ref: '#/components/responses/CSRFRejected'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/totp:
    post:
      operationId: verifyTOTP
      summary: Consume a one-time challenge and establish the administrator session
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPRequest'
      responses:
        '204':
          description: Two factors accepted; session and rotated CSRF cookies set, with no JWT response body
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '403':
          $ref: '#/components/responses/CSRFRejected'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/logout:
    post:
      operationId: logout
      summary: Revoke the current server-side session and clear cookies
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Session revoked; session and CSRF cookies cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CSRFRejected'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me:
    get:
      operationId: getCurrentAdministrator
      summary: Return the authenticated administrator identity
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    allocationServiceBearer:
      type: http
      scheme: bearer
...
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InvalidCredentials:
      description: Credentials, challenge, or TOTP are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: A valid, non-revoked session is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CSRFRejected:
      description: Same-origin or CSRF validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Account, client IP, or global authentication limit reached
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Request could not be processed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CSRFResponse:
      type: object
      additionalProperties: false
      required: [csrf_token]
      properties:
        csrf_token:
          type: string
          minLength: 32
    PasswordRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 256
        password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
    ChallengeResponse:
      type: object
      additionalProperties: false
      required: [challenge, expires_in]
      properties:
        challenge:
          type: string
          minLength: 32
        expires_in:
          type: integer
          enum: [120]
    TOTPRequest:
      type: object
      additionalProperties: false
      required: [challenge, code]
      properties:
        challenge:
          type: string
          minLength: 32
        code:
          type: string
          pattern: '^[0-9]{6}$'
          writeOnly: true
    MeResponse:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
        request_id:
          type: string
        run_id:
          type: string